Update to use nuxt pages
